Add display math support
