+++ package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "@fontsource/inter": "^4.5.12",
     "@fontsource/roboto": "^4.5.8",
+    "@khanacademy/simple-markdown": "^0.8.3",
     "@matrix-org/olm": "https://gitlab.matrix.org/api/v4/projects/27/packages/npm/@matrix-org/olm/-/@matrix-org/olm-3.2.12.tgz",
     "@tippyjs/react": "^4.2.6",
     "babel-polyfill": "^6.26.0",
@@ -42,7 +43,6 @@
     "react-google-recaptcha": "^2.1.0",
     "react-modal": "^3.15.1",
     "sanitize-html": "^2.7.1",
-    "simple-markdown": "^0.7.3",
     "tippy.js": "^6.3.7",
     "twemoji": "^14.0.2"
   },
